Move SkyModels to core
